Map generation tool
Make sure your maps and tile_types tables are empty.
To do this, you'll need to delete all characters and
then run this script.

If you want a different map, replace map.png or pass
the filename to the script. The scale is 1 pixel to
1 map tile.

Only 6 tile types are recognised - adjust the script
as required.